Use external KeycloakAuthInstance managing OIDC discovery
