+++ src/lib.rs
@@ -162,8 +162,8 @@ use role::Role;
 
 pub mod decode;
 pub mod error;
+pub mod instance;
 pub mod layer;
-pub mod middleware;
 pub mod oidc;
 pub mod oidc_discovery;
 pub mod role;
